requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
scikit-learn==1.3.2
pandas==2.2.2
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.0.0  # optional for LLM recommendations
